Bug Fix:   Still solving the memory over-write problem.  I think (this time)     its because I wasn't #including sys/time.h   Changed extensible->result to be a int instead of long
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk/net-snmp@143 06827809-a52a-0410-b366-d66718629ded

+++ agent/extensible/struct.h
@@ -8,7 +8,7 @@ struct extensible
    char name[STRMAX];
    char command[STRMAX];
    int type;
-   long result;
+   int result;
    char output[STRMAX];
    struct extensible *next;
    int miboid[MIBMAX];
